Create delete account page

+++ controllers/userController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require("express-async-handler");
 const User = require("../models/user");
 const Post = require("../models/post");
+const { body, validationResult } = require("express-validator");
 
 exports.user_get_posts = asyncHandler(async (req, res, next) => {
     const [user, userPosts] = await Promise.all([
@@ -11,14 +12,37 @@ exports.user_get_posts = asyncHandler(async (req, res, next) => {
 })
 
 exports.user_delete_get = (req, res, next) => {
-    res.render("user_delete");
+    if (req.user) {
+        if (req.params.id == req.user._id) {
+            res.render("user_delete", {title: "Delete Account", userid: req.user._id});
+        } else {
+            res.redirect("/");
+        }
+    } else {
+        res.redirect("/log-in");
+    }
 }
 
-exports.user_delete_post = asyncHandler(async (req, res, next) => {
-    await Promise.all([
-        Post.deleteMany({user: req.body.userid}),
-        User.findByIdAndRemove(req.body.userid),
-     ]) 
-    res.redirect("/");
-})
+exports.user_delete_post = [
+    body("delete", "Type DELETE to confirm")
+        .trim()
+        .equals("DELETE"),
+    
+    asyncHandler(async (req, res, next) => {
+        const errors = validationResult(req);
+
+        if (!errors.isEmpty()) {
+            res.render("user_delete", {title: "Delete Account", userid: req.user._id, errors: errors.array()});
+        } else {
+            await Promise.all([
+                Post.deleteMany({user: req.body.userid}),
+                User.findByIdAndRemove(req.body.userid),
+            ]) 
+            res.redirect("/");
+        }
+    })
+    
+]
+
+
